Replaced old navigation reducer and action creator
