+++ src/reducers/navigationReducer.js
@@ -1,25 +1,29 @@
-import { FETCH_NAVIGATION_SUCCESS, FETCH_NAVIGATION_REQUEST, FETCH_NAVIGATION_FAILURE } from "../actions/actionTypes";
-import initialState from "./initialState";
+// @flow
+const initialState: NavigationState = {
+  isFetching: false,
+  categories: [],
+};
 
-export default function navigationReducer(state = initialState.navigation, action) {
+export default function navigationReducer(state: NavigationState = initialState, action: Action): NavigationState {
   switch (action.type) {
-    case FETCH_NAVIGATION_REQUEST:
-      return { ...state, isFetching: true };
-    case FETCH_NAVIGATION_SUCCESS: {
-      // Filter out the data we need
-      const items = [];
-      action.navigation.forEach((element) => {
-        const item = {
-          id: element.id,
-          name: element.name,
-          items: element.object_list,
-        };
-        items.push(item);
-      });
-      return { ...state, items, isFetching: false };
+    case "FETCH_NAVIGATION_REQUEST":
+      return {
+        ...state,
+        isFetching: true,
+      };
+    case "FETCH_NAVIGATION_SUCCESS": {
+      const { categories } = action;
+      return {
+        ...state,
+        categories,
+        isFetching: false,
+      };
     }
-    case FETCH_NAVIGATION_FAILURE:
-      return { ...state, isFetching: false };
+    case "FETCH_NAVIGATION_FAILURE":
+      return {
+        ...state,
+        isFetching: false,
+      };
     default:
       return state;
   }
